requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.1
    # via llama-index
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via llama-index
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   nltk
dataclasses-json==0.6.3
    # via llama-index
deprecated==1.2.14
    # via llama-index
distro==1.9.0
    # via openai
dnspython==2.4.2
    # via pymongo
flask==3.0.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.12.2
    # via llama-index
greenlet==3.0.1
    # via
    #   playwright
    #   sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via
    #   llama-index
    #   openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
joblib==1.3.2
    # via nltk
llama-index==0.9.25.post1
    # via -r requirements.in
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.5.8
    # via llama-index
nltk==3.8.1
    # via llama-index
numpy==1.26.3
    # via
    #   llama-index
    #   pandas
openai==1.6.1
    # via llama-index
packaging==23.2
    # via
    #   marshmallow
    #   pytest
pandas==2.1.4
    # via
    #   -r requirements.in
    #   llama-index
playwright==1.40.0
    # via
    #   -r requirements.in
    #   playwright-stealth
playwright-stealth==1.0.6
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
pydantic==2.5.3
    # via openai
pydantic-core==2.14.6
    # via pydantic
pyee==11.0.1
    # via playwright
pymongo==4.6.1
    # via -r requirements.in
pytest==7.4.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
regex==2023.12.25
    # via
    #   nltk
    #   tiktoken
requests==2.31.0
    # via
    #   llama-index
    #   tiktoken
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy[asyncio]==2.0.25
    # via llama-index
tenacity==8.2.3
    # via llama-index
tiktoken==0.5.2
    # via llama-index
tqdm==4.66.1
    # via
    #   nltk
    #   openai
typing-extensions==4.9.0
    # via
    #   llama-index
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index
tzdata==2023.4
    # via pandas
urllib3==2.1.0
    # via requests
werkzeug==3.0.1
    # via flask
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via aiohttp
